+++ node/clients/search-segment/queries.ts
@@ -1,8 +1,0 @@
-export const searchSegment = `
-query searchSegment($userEmail: String, $isAuthenticated: Boolean $selectedFacets: [SelectedFacetsInput]) {
-  searchSegment(isAuthenticated: $isAuthenticated, userEmail: $userEmail, selectedFacets: $selectedFacets){
-    key
-    value
-  }
-}
-`